#12935: create framework folder and move internal files
