+++ tests/sweep_framework/query.py
@@ -7,7 +7,7 @@ import os
 import pathlib
 import pprint
 from elasticsearch import Elasticsearch, NotFoundError
-from tests.sweep_framework.statuses import TestStatus
+from tests.sweep_framework.framework.statuses import TestStatus
 from beautifultable import BeautifulTable, STYLE_COMPACT
 from termcolor import colored
 from elastic_config import *
